requirements.txt
fastapi
uvicorn
google-genai
sqlalchemy
psycopg2-binary
python-dotenv
google-adk